Fixed the Regression line
Fixed the regression line to plot on the scatterplot. Everything should
work now. Moved a ton to the plotData.py file.
